backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
textblob==0.17.1
gunicorn==21.2.0
python-dotenv==1.0.0
transformers==4.36.2
torch==2.2.1
nltk==3.8.1
langdetect==1.0.9
wordcloud==1.9.3
PyPDF2==3.0.1
python-docx==1.0.1
flask-limiter==3.5.0
flask-caching==2.1.0
flask-jwt-extended==4.6.0
emoji==2.10.1
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1
scikit-learn==1.3.2
